+++ components/authorpage/AuthorResultData.js
@@ -8,6 +8,7 @@ import AuthorBooks from "./AuthorBooks";
 import AuthorSeries from "./AuthorSeries";
 import Toast from "../global/Toast";
 import LibraryButton from "../global/LibraryButton";
+import { cleanImageUrl } from "../../utils/cleanImageUrl";
 
 const AuthorResultData = ({ scrapedData }) => {
   const [imageLoaded, setImageLoaded] = useState(false);
@@ -126,26 +127,23 @@ const AuthorResultData = ({ scrapedData }) => {
                   {/* Load WebP Image With JPG Fallback & 404 Not Found Image*/}
                   <picture className={imageLoaded ? "" : "hidden"}>
                     <source
-                      srcSet={`/img?url=${scrapedData.image
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&output=webp&maxage=30d`}
+                      srcSet={`/img?url=${cleanImageUrl(
+                        scrapedData.image
+                      )}&output=webp&maxage=30d`}
                       type="image/webp"
                       className="rounded-2xl mx-auto shadow-2xl drop-shadow-xl"
                     />
                     <source
-                      srcSet={`/img?url=${scrapedData.image
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&maxage=30d`}
+                      srcSet={`/img?url=${cleanImageUrl(
+                        scrapedData.image
+                      )}&maxage=30d`}
                       type="image/jpeg"
                       className="rounded-2xl mx-auto shadow-2xl drop-shadow-xl"
                     />
                     <img
-                      src={`/img?url=${scrapedData.image
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&maxage=30d`}
+                      src={`/img?url=${cleanImageUrl(
+                        scrapedData.image
+                      )}&maxage=30d`}
                       className="rounded-2xl border-2 mx-auto shadow-2xl drop-shadow-xl"
                       alt=""
                       width="286"