+++ components/librarypage/AuthorList.js
@@ -1,5 +1,7 @@
 /* eslint-disable @next/next/no-img-element */
 
+import { cleanImageUrl } from "../../utils/cleanImageUrl";
+
 const AuthorList = ({ libraryData }) => {
   return (
     <div
@@ -39,26 +41,23 @@ const AuthorList = ({ libraryData }) => {
                     {data.image && (
                       <picture>
                         <source
-                          srcSet={`/img?url=${data.image
-                            .replace("._SX50_SY75_", "")
-                            .replace("._SY75_", "")
-                            .replace("._SX50_", "")}&output=webp&maxage=30d`}
+                          srcSet={`/img?url=${cleanImageUrl(
+                            data.image
+                          )}&output=webp&maxage=30d`}
                           type="image/webp"
                           className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-900"
                         />
                         <source
-                          srcSet={`/img?url=${data.image
-                            .replace("._SX50_SY75_", "")
-                            .replace("._SY75_", "")
-                            .replace("._SX50_", "")}&maxage=30d`}
+                          srcSet={`/img?url=${cleanImageUrl(
+                            data.image
+                          )}&maxage=30d`}
                           type="image/jpeg"
                           className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-900"
                         />
                         <img
-                          src={`/img?url=${data.image
-                            .replace("._SX50_SY75_", "")
-                            .replace("._SY75_", "")
-                            .replace("._SX50_", "")}&maxage=30d`}
+                          src={`/img?url=${cleanImageUrl(
+                            data.image
+                          )}&maxage=30d`}
                           alt={`${data.name}'s image`}
                           width="98"
                           height="148"