+++ components/searchpage/BookResultData.js
@@ -1,6 +1,7 @@
 /* eslint-disable @next/next/no-img-element */
 import Link from "next/link";
 import Meta from "../global/Meta";
+import { cleanImageUrl } from "../../utils/cleanImageUrl";
 
 const SearchResults = (props) => {
   const capitalizeTitle = (title) => {
@@ -70,26 +71,21 @@ const SearchResults = (props) => {
                 <div className="flex mr-8">
                   <picture>
                     <source
-                      srcSet={`/img?url=${data.cover
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&output=webp&maxage=30d`}
+                      srcSet={`/img?url=${cleanImageUrl(
+                        data.cover
+                      )}&output=webp&maxage=30d`}
                       type="image/webp"
                       className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-800"
                     />
                     <source
-                      srcSet={`/img?url=${data.cover
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&maxage=30d`}
+                      srcSet={`/img?url=${cleanImageUrl(
+                        data.cover
+                      )}&maxage=30d`}
                       type="image/jpeg"
                       className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-800"
                     />
                     <img
-                      src={`/img?url=${data.cover
-                        .replace("._SX50_SY75_", "")
-                        .replace("._SY75_", "")
-                        .replace("._SX50_", "")}&maxage=30d`}
+                      src={`/img?url=${cleanImageUrl(data.cover)}&maxage=30d`}
                       alt={`${data.title} book cover`}
                       width="98"
                       height="148"