+++ components/seriespage/SeriesList.js
@@ -1,5 +1,6 @@
 /* eslint-disable @next/next/no-img-element */
 import Link from "next/link";
+import { cleanImageUrl } from "../../utils/cleanImageUrl";
 
 const SeriesList = (props) => {
   return (
@@ -60,29 +61,23 @@ const SeriesList = (props) => {
                   )}
                 </div>
               </div>
+
               <div className="flex mr-8">
                 <picture>
                   <source
-                    srcSet={`/img?url=${data.cover
-                      .replace("._SX120_SY180_", "")
-                      .replace("._SX120_", "")
-                      .replace("._SY180_", "")}&output=webp&maxage=30d`}
+                    srcSet={`/img?url=${cleanImageUrl(
+                      data.cover
+                    )}&output=webp&maxage=30d`}
                     type="image/webp"
                     className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-900"
                   />
                   <source
-                    srcSet={`/img?url=${data.cover
-                      .replace("._SX120_SY180_", "")
-                      .replace("._SX120_", "")
-                      .replace("._SY180_", "")}&maxage=30d`}
+                    srcSet={`/img?url=${cleanImageUrl(data.cover)}&maxage=30d`}
                     type="image/jpeg"
                     className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-900"
                   />
                   <img
-                    src={`/img?url=${data.cover
-                      .replace("._SX120_SY180_", "")
-                      .replace("._SX120_", "")
-                      .replace("._SY180_", "")}&maxage=30d`}
+                    src={`/img?url=${cleanImageUrl(data.cover)}&maxage=30d`}
                     alt={`${data.title} book cover`}
                     width="98"
                     height="148"
@@ -157,17 +152,19 @@ const SeriesList = (props) => {
               <div className="flex mr-8">
                 <picture>
                   <source
-                    srcSet={`/img?url=${data.cover}&output=webp&maxage=30d`}
+                    srcSet={`/img?url=${cleanImageUrl(
+                      data.cover
+                    )}&output=webp&maxage=30d`}
                     type="image/webp"
                     className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-800"
                   />
                   <source
-                    srcSet={`/img?url=${data.cover}&maxage=30d`}
+                    srcSet={`/img?url=${cleanImageUrl(data.cover)}&maxage=30d`}
                     type="image/jpeg"
                     className="rounded-lg shadow-sm drop-shadow-sm bg-white dark:bg-slate-800"
                   />
                   <img
-                    src={`/img?url=${data.cover}&maxage=30d`}
+                    src={`/img?url=${cleanImageUrl(data.cover)}&maxage=30d`}
                     alt={`${data.title} book cover`}
                     width="98"
                     height="148"