+++ pages/book/show/[...slug].js
@@ -43,7 +43,7 @@ const Slug = () => {
   }, [slug]);
 
   return (
-    <div>
+    <>
       <div className="bg-gradient-to-tr from-rose-50 to-rose-200 dark:bg-gradientedge text-gray-900 dark:text-gray-100 min-h-screen">
         <Header />
         {/* Show loader or error based on the scraper data response */}
@@ -73,7 +73,7 @@ const Slug = () => {
         )}
         <Footer />
       </div>
-    </div>
+    </>
   );
 };